Fix how field types are denoted
Before adding field type support prior data in InfluxDB was defaulting
to type float if the type was of a numeric value. After adding initial
support of different vale types, the `to_line_protocol` function for a
point would default an Elixir integer type to a Influx integer type.
However, these would cause a bug in existing InfluxDBs that use this
library.

This allows the caller to explicitly set the type of a field by
providing a tuple of `{field_value, field_type}`.

This also provides a `:type` option to `InfluxEx.Point.add_field/2`
to allow explicitly setting the type.

When using `InfluxEx.Point.add_fields/2` the default numeric type will
be float unless specified.